refactor: move the TraceResults type to homeboy-extension-contract
The finale of the trace subsystem breakdown. With both blockers cleared
(TracePreviewMetadata #8841, RigStateSnapshot #8843) and all 14+ parsing types
already moved, TraceResults - the top-level trace aggregate - moves wholesale to
the contract crate.

Every field type now resolves from contract/leaf crates: the trace parsing types
(TraceArtifact/Assertion/provenance/etc.), the span/event aliases (observation
timeline), TracePreviewMetadata, and RigStateSnapshot. The trace parse functions
(using error/structured_sidecar/Path) stay in core.

This collapses the trace reverse-edge surface from ~45 references to 3 core
files - mirroring the bench cluster's 58 -> 4. Core's parsing.rs re-exports
TraceResults so all consumers stay stable.
