+++ crates/homeboy-extension-contract/src/lib.rs
@@ -65,6 +65,7 @@ pub mod source_metadata_repair;
 pub mod test_drift;
 pub mod trace_config;
 pub mod trace_preview;
+pub mod trace_results;
 pub mod update_output;
 pub mod version;
